Move files into their own package under feature flags
