+++ app/src/androidTest/java/com/hedvig/app/feature/referrals/tab/SwipeToRefreshNewDataTest.kt
@@ -13,7 +13,7 @@ import com.hedvig.app.util.FeatureFlagRule
 import com.hedvig.app.util.LazyActivityScenarioRule
 import com.hedvig.app.util.apolloResponse
 import com.hedvig.app.util.context
-import com.hedvig.app.util.featureflags.Feature
+import com.hedvig.app.util.featureflags.flags.Feature
 import com.hedvig.app.util.swipeDownInCenter
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
 import io.github.kakaocup.kakao.screen.Screen
