+++ app/src/androidTest/java/com/hedvig/app/util/FeatureFlagRule.kt
@@ -1,8 +1,8 @@
 package com.hedvig.app.util
 
 import com.hedvig.app.featureManagerModule
-import com.hedvig.app.util.featureflags.Feature
 import com.hedvig.app.util.featureflags.FeatureManager
+import com.hedvig.app.util.featureflags.flags.Feature
 import io.mockk.coEvery
 import io.mockk.mockk
 import org.junit.rules.ExternalResource
@@ -13,8 +13,9 @@ import org.koin.dsl.module
 class FeatureFlagRule(
     vararg flags: Pair<Feature, Boolean>,
 ) : ExternalResource() {
-    private val mockModule = module {
-        single {
+    @Suppress("RemoveExplicitTypeArguments")
+    private val mockFeatureManagerModule = module {
+        single<FeatureManager> {
             val mock = mockk<FeatureManager>()
             flags.forEach { coEvery { mock.isFeatureEnabled(it.first) } returns it.second }
             mock
@@ -23,11 +24,11 @@ class FeatureFlagRule(
 
     override fun before() {
         unloadKoinModules(featureManagerModule)
-        loadKoinModules(mockModule)
+        loadKoinModules(mockFeatureManagerModule)
     }
 
     override fun after() {
-        unloadKoinModules(mockModule)
+        unloadKoinModules(mockFeatureManagerModule)
         loadKoinModules(featureManagerModule)
     }
 }
