+++ app/src/main/java/com/hedvig/app/feature/home/model/HomeItemsBuilder.kt
@@ -8,8 +8,8 @@ import com.hedvig.app.R
 import com.hedvig.app.feature.claims.ui.commonclaim.CommonClaimsData
 import com.hedvig.app.feature.claims.ui.commonclaim.EmergencyData
 import com.hedvig.app.feature.home.ui.claimstatus.data.ClaimStatusCardUiState
-import com.hedvig.app.util.featureflags.Feature
 import com.hedvig.app.util.featureflags.FeatureManager
+import com.hedvig.app.util.featureflags.flags.Feature
 
 @OptIn(ExperimentalStdlibApi::class)
 class HomeItemsBuilder(
