+++ app/src/main/java/com/hedvig/app/feature/profile/ui/tab/ProfileFragment.kt
@@ -19,7 +19,6 @@ import com.hedvig.app.feature.profile.ui.myinfo.MyInfoActivity
 import com.hedvig.app.feature.profile.ui.payment.PaymentActivity
 import com.hedvig.app.feature.settings.MarketManager
 import com.hedvig.app.feature.settings.SettingsActivity
-import com.hedvig.app.featureflags.FeatureFlagEntryProvider
 import com.hedvig.app.util.apollo.format
 import com.hedvig.app.util.apollo.toMonetaryAmount
 import com.hedvig.app.util.extensions.showAlert
@@ -39,7 +38,6 @@ class ProfileFragment : Fragment(R.layout.profile_fragment) {
     private val loggedInViewModel: LoggedInViewModel by sharedViewModel()
     private var scroll = 0
     private val marketManager: MarketManager by inject()
-    private val featureFlagEntryProvider: FeatureFlagEntryProvider by inject()
 
     override fun onResume() {
         super.onResume()
@@ -130,7 +128,6 @@ class ProfileFragment : Fragment(R.layout.profile_fragment) {
                                 ) {
                                     startActivity(Intent(requireContext(), AboutAppActivity::class.java))
                                 },
-                                featureFlagEntryProvider.addEntry(requireContext()),
                                 ProfileModel.Logout,
                             )
                         )
