+++ app/src/main/java/com/hedvig/app/service/badge/ReferralsNotificationBadgeService.kt
@@ -1,7 +1,7 @@
 package com.hedvig.app.service.badge
 
-import com.hedvig.app.util.featureflags.Feature
 import com.hedvig.app.util.featureflags.FeatureManager
+import com.hedvig.app.util.featureflags.flags.Feature
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emitAll
 import kotlinx.coroutines.flow.flow
