+++ app/src/main/java/com/hedvig/app/util/featureflags/flags/FeatureFlagProvider.kt
@@ -1,0 +1,5 @@
+package com.hedvig.app.util.featureflags.flags
+
+interface FeatureFlagProvider {
+    suspend fun isFeatureEnabled(feature: Feature): Boolean
+}
