--- app/src/release/java/com/hedvig/app/featureflags/FeatureFlagEntryProvider.kt
@@ -1,8 +1,0 @@
-package com.hedvig.app.featureflags
-
-import android.content.Context
-import com.hedvig.app.feature.profile.ui.tab.ProfileModel
-
-class FeatureFlagEntryProvider {
-    fun addEntry(context: Context): ProfileModel.Row? = null
-}
